Renamed the Comment class to SVGComment due to a conflict with the AIFF header. Changed some ints to NSIntegers. Create functions NSStringFromCGRect and NSStringFromCGPoint for OS X. Have the black color in SVGTextElement be a macro that works on OS X and iOS. Link to the Quartz framework and livxml

+++ Source/DOM classes/Core DOM/NamedNodeMap.m
@@ -86,7 +86,7 @@
 
 -(unsigned long)length
 {
-	int count = [self.internalDictionary count];
+	NSUInteger count = [self.internalDictionary count];
 	
 	for( NSDictionary* namespaceDict in self.internalDictionaryOfNamespaces )
 	{
